requirements.txt
phi 
groq
exa_py
python-dotenv